Rename removeEvents.md to removeEvents_Proposal.md
In this piece of documentation about removeEvents, it looks as though the original text was typed with a word-processor implementing 'smart quotes'. I have just changed these to standard quotes: if a user copies the method signature 'as is' and pastes it into his code editor, an error will be generated — which will not be easy to correct if the font-size of the editor window is small. The good practice about standard quotes should be applied throughout the documentation documents.