Include `didMount` and `willUnmount` in documented example.
Descriptions of the custom views with JavaScript should at least include a simple reference to the `didMount` and `willUnmount` methods that are invoked when the custom view has just loaded or is about to be unloaded, respectively. I had to go digging in the test cases to find that these event existed and they seem a common use case.