Prefer arrow function over .bind(this)
In my personal opinion, it's better to use an arrow function as a callback then directly passing in a function and using .bind(this) to keep the scope.
Arrow functions are now widely know and provide more flexability, plus arrow functions do not have a scope.

When the function is called by the arrow function, the other function will still remain in the correct scope aswell, so no need for custom binding